+++ apps/ymir-sdl3/src/app/ui/widgets/input_widgets.hpp
@@ -27,7 +27,7 @@ private:
 
     void CaptureButton(input::InputBind &bind, size_t elementIndex, void *context);
     void CaptureTrigger(input::InputBind &bind, size_t elementIndex, void *context);
-    void CaptureAxis1D(input::InputBind &bind, size_t elementIndex, void *context);
+    void CaptureAxis1D(input::InputBind &bind, size_t elementIndex, void *context, bool bipolar);
     void CaptureAxis2D(input::InputBind &bind, size_t elementIndex, void *context);
 
     void MakeDirty();